arm: tegra: Add the Skidata carrier board
Re-build the skidata carrier board from the patch in the
tamonten/l4t-experimental-r16 branch. This new version use the generic
Tamonten COM support recently added.

It also rename the board from 'skidata' to 'skidata-carrier' in order
to avoid confusion if another Skidata board needs be supported in the
future.

The ADC device, and its regulator, have been removed as the ADC was
only on early version of this board.

Change-Id: I0b559d8a6ef52949d95f73f06b54f5719a2a67b9
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/445
Reviewed-by: Dirk Leber <dirk.leber@avionic-design.de>