rtc: pcf8523: Clear RTC if time is unreliable
The pcf8523 has a flag to signal that the oscillator stopped at some point
since the flag has been last cleared. This can happen if the voltage drops
below a minimum operating threshold but keeps above the overall minimum
voltage. In this case the time registers are not incremented anymore and hence
the time freezes. Additionally the flag is set on initial power up of the
chip, so that it can be reliably used to detect possibly bad values.

While the driver already checked the flags value it did not handle it further
except for printing a debug message, if enabled. To make some better use of
this feature clear the time registers to zero in the error case, which allows
to detect the error case from userspace.

Change-Id: Iceb98dc7b4dea270461da1019bfa9c445b1e1c67
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/2080
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>